+++ src/main/java/tornadofx/FX.kt
@@ -497,6 +497,28 @@ fun EventTarget.addChildIfPossible(node: Node, index: Int? = null) {
 }
 
 /**
+ * Bind the children of this Layout node to the given observable list of items by converting
+ * them into nodes via the given converter function. Changes to the source list will be reflected
+ * in the children list of this layout node.
+ */
+inline fun <reified T> EventTarget.bindChildren(sourceList: ObservableList<T>, noinline converter: (T) -> Node): ListConversionListener<T, Node> {
+    val children = getChildList() ?: throw IllegalArgumentException("Unable to extract child nodes from $this")
+    return children.bind(sourceList, converter)
+}
+
+/**
+ * Bind the children of this Layout node to the given observable list of items by converting
+ * them into UIComponents via the given converter function. Changes to the source list will be reflected
+ * in the children list of this layout node.
+ */
+inline fun <reified T> EventTarget.bindComponents(sourceList: ObservableList<T>, noinline converter: (T) -> UIComponent): ListConversionListener<T, Node> {
+    val children: MutableList<Node> = getChildList() ?: throw IllegalArgumentException("Unable to extract child nodes from $this")
+    val componentConverter: (T) -> Node = { converter(it).root }
+    return children.bind(sourceList, componentConverter)
+}
+
+
+/**
  * Find the list of children from a Parent node. Gleaned code from ControlsFX for this.
  */
 fun EventTarget.getChildList(): MutableList<Node>? = when (this) {